Added in Sprites Added in Appear/DisappearLogic.
Still some bugs going on,  Ok and bunny shouldnt register while round is in play /// there is no agreement.

Need to fix those holes in the logic and add logic that checks for correct answer and adds to tally.

Then I really gotta buckle down and work on multiplayer and co-location.